% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{chop_trials}
\alias{chop_trials}
\title{chop_trials: Chop continuous recording into trials}
\usage{
chop_trials(df, start_postfix = ".begin", end_postfix = ".end")
}
\arguments{
\item{df}{}
}
\description{
chop_trials: Chop continuous recording into trials
}
