% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_csv.R
\name{clean_csv_names}
\alias{clean_csv_names}
\title{clean_csv_names: Clean up column names}
\usage{
clean_csv_names(names)
}
\arguments{
\item{names}{Column names to clean}
}
\description{
clean_csv_names: Clean up column names
}
