% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_csv.R
\name{clean_names}
\alias{clean_names}
\title{clean_names: Clean up column names}
\usage{
clean_names(names)
}
\arguments{
\item{names}{Column names to clean}
}
\description{
clean_names: Clean up column names
}
