% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_csv.R
\name{csv_col_defs}
\alias{csv_col_defs}
\title{csv_col_defs: Column definitions for Neon exported csv files}
\usage{
csv_col_defs(name)
}
\arguments{
\item{name}{One of "gaze", "3d_eye_states", "blinks", "events", "fixations",
"imu", "world_timestamps"}
}
\value{
readr-compatible cols() definition
}
\description{
csv_col_defs: Column definitions for Neon exported csv files
}
