% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonR-package.R
\docType{package}
\name{neonR-package}
\alias{neonR}
\alias{neonR-package}
\title{neonR: Read and organize data from Pupil Labs Neon}
\description{
Read and organize data from Pupil Labs Neon.
}
\author{
\strong{Maintainer}: Brian Lau \email{brian.lau@upmc.fr} (\href{https://orcid.org/0000-0001-8615-8244}{ORCID})

}
\keyword{internal}
