% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.neonr}
\alias{print.neonr}
\title{#' @export
print <- function(x, ...){
UseMethod("print", x)
}
print.neonr:}
\usage{
\method{print}{neonr}(x)
}
\arguments{
\item{obj}{}
}
\description{
#' @export
print <- function(x, ...){
UseMethod("print", x)
}
print.neonr:
}
