% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_csv.R
\name{read_csv_file}
\alias{read_csv_file}
\title{read_csv_file: Read a single Neon exported csv file}
\usage{
read_csv_file(filename, convert_time_cols = T, clean_names = T)
}
\arguments{
\item{convert_time_cols}{Convert columns with timestamps to POSIXct (default TRUE)}

\item{clean_names}{Clean column names (default TRUE)}
}
\description{
read_csv_file: Read a single Neon exported csv file
}
