% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_csv.R
\name{read_exported_dir}
\alias{read_exported_dir}
\title{read_exported_dir: Read all Neon exported csv files in directory}
\usage{
read_exported_dir(path = ".", ...)
}
\arguments{
\item{...}{}
}
\description{
read_exported_dir: Read all Neon exported csv files in directory
}
