% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{chop_world_video_trials}
\alias{chop_world_video_trials}
\title{chop_world_video_trials: Chop world (scene) video into trials}
\usage{
chop_world_video_trials(
  obj,
  videofile,
  save_dir = fs::path_dir(videofile),
  trial = NULL,
  overwrite = T,
  silent = T
)
}
\arguments{
\item{videofile}{Name of video file}

\item{trial}{TODO}

\item{overwrite}{Overwrite existing files (default = TRUE)}

\item{silent}{Ignore stdout of ffmpeg call (default = TRUE)}

\item{df}{A trial-based neon_data object}
}
\description{
chop_world_video_trials: Chop world (scene) video into trials
}
